sockbuf policy: back to EPOLLONESHOT
Return back to EPOLLONESHOT and loop in the sockmgr::process_regular.
Epoll return few events in one action (normal data and FIN---connection
close), so do ::read before it return -1 and set EAGAIN (or EINTR) or show
0 as close indicator.

Don't do epoll in write: when simultaneous wait in epoll and poll,
poll may not be alarmed. Try ::write while it return -1 and
set EAGAIN (or EINTR) (with increasing delay)