Janus: total order, check leader
Check bool value instead of compare leader_ with self_id():
not only avoid 16 bytes compare, but avoid mutex in self_id().

In most cases we need only whether this node is a group leader
or not, and, may be, only for (next) leader election
we ask for current leader address.