yard: table of sections
Introduce table of sections.
Table of sections is records of pairs:

  <section number> <section offset>

The last entry is (0,0). Now only section number
significant---0 is end of sections table.

In this table not only sections offsets, but
some info too: id 3 is number of buskets
in stored hash table.

All offsets in section should be relative section
offset---not sure that this is corrected now.