sockios: fix busy loop bug
Busy loop in connect_processor worker thread
appers if clients sockstream status is not good
and data is available in buffer. In this case we
get infinite loop in process_request.

Stream status must be also checked, may be it
is better to make users connect function return
boolean value to indicate error status.