stem: fix memory leack in NetTransport
NetTransport objects were created with some
string description, so pair (id,string) was
added to iheap for each NetTransport objects.
The problem is that NetTransport object by itself
is never enabled and so (id, obj) pair never goes
at heap container, so it is never Unsubcribed.
This leads that ids of NetTransport object are never
freed from iheap container and its size  always grow
which result in memory leack and poor perfomance.

This patch is quick and dirty - it simply removes useless
string description for NetTransport, but the problem is
deeper - object and it's desctipion must be added to
EvManager tables in one step.