yard: common ancestor for two commits
Implementation of search of common ancestor
for two commits.

Commits is DAG (but not tree). This algo is
like search in breadth for left and right
nodes. We traverse transposed graph 'in breadth'.
Ancestors of left colored black, ancestors
of right colored gray. When we see non-white
color of vertice, then this vertice is common
ancestor, nearest in some terms.

common_ancestor(2,3)

1.

1   2   3  4
 \/  \   \/
  5   6  7
   \ /  /
    8  /
    \ /
     9
     |
   root

2.

1  2b  3g 4
 \/  \   \/
  5   6  7
   \ /  /
    8  /
    \ /
     9
     |
    root

3.

1  2b   3g 4
 \/  \   \/
  5b  6b 7g
   \ /  /
    8  /
    \ /
     9
     |
   root

4.

1  2b   3g 4
 \/  \   \/
  5b  6b 7g
   \ /  /
    8b /
    \ /
     9g
     |
   root

5. ok, 9 is common ancestor of 2 and 3.

Note, that this algo don't give 'perfect' common ancestor
(between successors of common ancestor no another
common ancestor), but enough for me.