Added a first version of key iterator.
Key iterator used to iterate through a range of keys in the block.
Now it is only constant.
It is bidirectional input (!) iterator.
It CANNOT be an output iterator because an entry may not contain a space
for any key.

Also it has a method to extract an entry which contains a specified
key.