Modified key iterators.
Extracted traversal logic of the iterator to a separate iterator file.
Added const_key_iterator and key_iterator.
Added a method to get a entry by a key iterator.

Therefore a key iterator is an not just a iterator through keys in the
block, but a coordinate in the block.

Known problems:
- There are some overhead associated with entry size and entry type.
- Ugly code in the implementation of key_begin and key_end functions.
- There are two insert functions.
- One can invoke route function for a data block (this is senseless)
- One can invoke lookup function for a index block (this is senseless)
- for its support random-access traversal
  key_iterator and const_key_iterator have input_iterator_tag.
  (Therefore I cannot use a binary search in the lookup and route
  function implementation)