Made BTree more robust.
When the block of the tree is overfilled it divides.
And now two new blocks are appended to the file
(So the old version of the block is still in the file).
and the delimiter is added to the parent block.

If the block is not overfilled it replaces the original.

The thing is to change (!) only one block at a time.
So the insertion is now a transaction.

Note:
I don't control the OS IO buffering.

And it may occur the situation in which the parent block is replaced but
the children is not appended. So we need a way to distinguish
an empty block (which contains garbage) and a valid block.