Redesigned BTree.
1)
Now the interface of the BTree is the following:

coordinate_type lookup(const key_type& key);
coordinate_type lookup(const coordinate_type& start, const key_type&
key);
coordinate_type insert(coordinate_type path, const key_type& key, const
value_type& value);

`lookup` find the block where `key` can be found/inserted.

The second version of `lookup` is just an enhancement of the first one.
It can be much faster if `start` is close to the result of the lookup.
But it (MUST) does the same amount of IOs otherwise.

`insert` function inserts a (key, value) to the specified block.
If the overflow doen't occur the return value MUST be the coordinate of
the specified block (i.e. be equal to `path`)

Return value specifies an affected subtree.

Note. It invalidates the `path`.
Note. Empty coordinate structure means whole tree.

2)
Also the interface of the `block_type` is changed.

Removed the non-const version of the find function.
Because you can't change the elements of the `block_type` without
violation of an estimate of a size of serilized `block_type`.

Added a erase function.

3)
Added a `get_block` function.
Please, do not use a cache_ directly!

Note. Now cache_ is consistent with an information on a disk.