Added a drop_cache test
This test actually doestn't do much.
It just writes a big file. That's all. Why?
There are two reasons.

First of all it is a good idea to measure raw IO
before trying to interpret different yard perfomance tests

The second reason is that it is the most simplest way to
drop filesystem caches.