sockios: SOCK_NONBLOCK and SOCK_CLOEXEC
Linux has non-standard options that may be or'ed
with socket type: SOCK_NONBLOCK and SOCK_CLOEXEC.

It atomically set close-on-exec flag for the
new descriptor(s) and atomically mark descriptor(s) as
non-blocking, respective.

If this flags are present, use ones, and avoid
fcntl calls to set O_NONBLOCK flag, when possible.